Fixed issue where listing poster was able to delete listings that were in progress

+++ listings/views.py
@@ -119,18 +119,21 @@ def modify_listings(request, pk):
 @login_required
 def delete_listing(request, pk):
     item = Listings.objects.get(id=pk)
-    if request.user == item.huntee:
-        if request.method == "POST":
-            item.delete()
+    if request.user == item.huntee and item.status == 'Strutting' or request.user == item.huntee and item.status == 'Payment Issued':
+        if request.user == item.huntee:
+            if request.method == "POST":
+                item.delete()
+                return redirect('current_listings')
+        else:
             return redirect('current_listings')
+        unread_messages = MessagingService.get_unread_messages(request, request.user)
+        has_unread_messages = True
+        num_of_unread_messages = unread_messages.count()
+        if not unread_messages:
+            has_unread_messages = False
+        return render(request, 'listings/delete_listing.html', {'item':item, 'has_unread_messages':has_unread_messages, 'num_of_unread_messages':num_of_unread_messages})
     else:
         return redirect('current_listings')
-    unread_messages = MessagingService.get_unread_messages(request, request.user)
-    has_unread_messages = True
-    num_of_unread_messages = unread_messages.count()
-    if not unread_messages:
-        has_unread_messages = False
-    return render(request, 'listings/delete_listing.html', {'item':item, 'has_unread_messages':has_unread_messages, 'num_of_unread_messages':num_of_unread_messages})
 
 @login_required
 def return_listing(request, pk):
